Update to Menu Screen
Overhaul to the design of the menu screen. Currently in the process of
making a 'register username' form. Now working with browser cookies.
Accounts are not universal and will not be remembered by the server
(outside of the game info, which will be stored on a database for a
month or so).